refactor: split files, add documentation
